requirements.txt
Django>=6.0,<7.0
python-dotenv>=1.0,<2.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.29,<1.0
